+++ packages/xrpl/src/models/transactions/common.ts
@@ -386,6 +386,10 @@ export interface BaseTransaction extends Record<string, unknown> {
    * The network id of the transaction.
    */
   NetworkID?: number
+  /**
+   * The delegate account that is sending the transaction.
+   */
+  Delegate?: Account
 }
 
 /**
@@ -396,6 +400,7 @@ export interface BaseTransaction extends Record<string, unknown> {
  * @param common - An interface w/ common transaction fields.
  * @throws When the common param is malformed.
  */
+// eslint-disable-next-line max-statements, max-lines-per-function -- lines required for validation
 export function validateBaseTransaction(
   common: Record<string, unknown>,
 ): asserts common is BaseTransaction {
@@ -448,6 +453,15 @@ export function validateBaseTransaction(
   validateOptionalField(common, 'TxnSignature', isString)
 
   validateOptionalField(common, 'NetworkID', isNumber)
+
+  validateOptionalField(common, 'Delegate', isAccount)
+
+  const delegate = common.Delegate
+  if (delegate != null && delegate === common.Account) {
+    throw new ValidationError(
+      'BaseTransaction: Account and Delegate addresses cannot be the same',
+    )
+  }
 }
 
 /**