+++ packages/xrpl/HISTORY.md
@@ -6,6 +6,7 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 
 ### Added
 * Support for `NFTokenMintOffer` (XLS-52)
+* Support for `Account Permissions` and `Account Permission Delegation` (XLS-74d, XLS-75d)
 * Add support for `Batch` amendment (XLS-56)
 
 ### Fixed